- Added an ESlint tool to notify me of errors so i could reduce the time debuging wih codacy since codcy uses the same linter - Modified some linting rule to make it work with my spacing choice - Updated the readme to display the results of Codacy
Just deployed to heroku
- required a lot of commits as i was having issues
- Updated my readme to reflect the changes and show my latest progress
- FIxed #5
