+++ .babelrc
@@ -1,3 +1,3 @@
 {
-	"presets": ["env", "stage-0"]
+	"presets": ["env"]
 }
