+++ .eslintrc.json
@@ -1,0 +1,23 @@
+{
+    "env": {
+        "browser": true,
+        "es2020": true
+    },
+    "extends": [
+        "standard"
+    ],
+    "parserOptions": {
+        "ecmaVersion": 11,
+        "sourceType": "module"
+    },
+    "rules": {
+        "indent": "off",
+        "no-tabs": "off",
+        "semi-spacing": "warn",
+        "comma-spacing": "warn",
+        "semi": [
+            "error",
+            "always"
+        ]
+    }
+}
