+++ README.md
@@ -1,3 +1,23 @@
 # Quick mechanic
 
-Team-291-Backend
+Team-291-Backend builds a system for a getting a mechanic when you car breakdowns
+
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/0f431f9d67854c3ea59f3010d7f7b383)](https://app.codacy.com/gh/BuildForSDGCohort2/Team-291-Backend?utm_source=github.com&utm_medium=referral&utm_content=BuildForSDGCohort2/Team-291-Backend&utm_campaign=Badge_Grade_Settings)
+
+## Installation Steps
+
+```sh
+git clone https://github.com/Dru-Go/Team-291-Backend
+
+npm i
+
+npm run dev # Runs the Enviroment in dev mode
+
+# Go to localhost:3000/graphql to view the graphql server
+
+npm run build # Builds the server for production
+```
+
+## Deployed to heroku
+
+Go to [Link](https://quick-mechanic.herokuapp.com/graphql) to view the deployed graphql server
