+++ package.json
@@ -1,11 +1,19 @@
 {
 	"name": "Team-291-Backend",
 	"version": "1.0.0",
+	"engines": {
+		"node": "12.18.3",
+		"npm": "6.14.6"
+	},
 	"description": "[![Codacy Badge](https://api.codacy.com/project/badge/Grade/0f431f9d67854c3ea59f3010d7f7b383)](https://app.codacy.com/gh/BuildForSDGCohort2/Team-291-Backend?utm_source=github.com&utm_medium=referral&utm_content=BuildForSDGCohort2/Team-291-Backend&utm_campaign=Badge_Grade_Settings)",
 	"main": "index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "nodemon ./index.js --exec babel-node -e js"
+		"dev": "nodemon src/index.js --exec babel-node -e js",
+		"clean": "rm -rf build && mkdir build",
+		"build-babel": "babel -d ./build ./src -s",
+		"build": "npm run clean && npm run build-babel",
+		"start": "npm run build && node ./build/index.js"
 	},
 	"repository": {
 		"type": "git",
@@ -19,14 +27,24 @@
 	},
 	"homepage": "https://github.com/Dru-Go/Team-291-Backend#readme",
 	"devDependencies": {
+		"@babel/core": "^7.11.4",
 		"babel-cli": "^6.26.0",
 		"babel-preset-env": "^1.7.0",
-		"babel-preset-stage-0": "^6.24.1"
+		"babel-preset-stage-0": "^6.24.1",
+		"eslint": "^7.7.0",
+		"eslint-config-standard": "^14.1.1",
+		"eslint-plugin-import": "^2.22.0",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-promise": "^4.2.1",
+		"eslint-plugin-standard": "^4.0.1"
 	},
 	"dependencies": {
+		"cors": "^2.8.5",
+		"dotenv": "^8.2.0",
 		"express": "^4.17.1",
 		"express-graphql": "^0.9.0",
 		"graphql": "^15.3.0",
-		"nodemon": "^2.0.4"
+		"nodemon": "^2.0.4",
+		"babel-preset-env": "^1.7.0"
 	}
 }
