+++ src/resolver.js
@@ -1,0 +1,7 @@
+// ANCHOR we define the resolvers which are used as routes and controllers
+const root = {
+    hello: () => 'Hello Worlds'
+    // NOTE Code goes here...
+};
+
+export default root;
