+++ src/schema.js
@@ -1,7 +1,7 @@
 // ANCHOR we define the schema which are used to structure the data flow b/n routes
-import { buildSchema } from "graphql";
+import { buildSchema } from 'graphql';
 
-const schema = buildSchema(`
+const schema = buildSchema(` 
 type Query {
 	hello: String
 }
